requirements-phase3.txt
sqlalchemy==2.0.23
sqlalchemy-utils==0.41.1
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
aiohttp==3.9.1
jose-cryptography==1.0.0
pydantic-settings==2.1.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-httpx==0.26.0
